requirements.txt
langchain>=0.3.0
python-dotenv==1.0.0
pydantic==2.10.6
SQLAlchemy==1.4.50
python-docx==1.1.0
asyncpg==0.28.0
pgvector==0.2.4
langgraph>=0.2
langchain_gigachat==0.3.3
gigachat==0.1.37.post1
typing-extensions>=4.5.0
